--- docs/sdlc-harness/research/strict-agent-evaluation-network-isolation.md
@@ -1,130 +1,0 @@
-# 断网不只是安全策略：让 Coding Agent 评测结果可解释
-
-> 本文记录 BitFun 在 Harbor / SWE-bench Pro 上构建严格评测环境时的取舍。它讨论的是**公开、历史 issue 构成的离线评测**，而不是一般开发环境的联网策略。
-
-## 摘要
-
-如果 agent 可以在运行时访问互联网，或读取目标仓库的未来 Git 历史，那么一个 `reward = 1` 不能单独说明它理解了本地代码、定位了问题并独立写出了修复。它也可能说明 agent 找到了已经公开的 PR、commit、隐藏测试或 gold patch。对于由历史公开仓库构成的 SWE-bench 类评测，这会把“检索到答案”的能力与“解决问题”的能力混在同一个分数里。
-
-因此，严格评测需要同时控制两类渠道：
-
-1. 在 agent 执行阶段阻断除模型服务外的网络出口；
-2. 在 agent 启动前移除原始 Git 历史，并初始化为单提交仓库。
-
-这不是为了让真实开发中的 agent 失去正常工具能力。相反，它是为了让离线基准的分数有明确含义：在给定的问题描述、当前工作区和允许的工具下，agent 是否能自行完成修复。
-
-## 为什么非断网结果不可信
-
-这里的“不可信”不是说每一个通过的 patch 都一定抄袭，也不是说联网 agent 在真实开发中不应存在。它的准确含义是：**在没有逐条轨迹审计时，分数无法可靠归因到本地推理和实现能力。**
-
-历史公开 issue 有一个特别的结构性问题：任务所对应的修复通常已经在上游合并。只要 agent 能访问 GitHub、代码搜索服务、镜像站或 Git 历史，它就有机会先找到答案，再把答案复写到工作区。常见路径包括：
-
-| 泄漏通道 | 典型动作 | 为什么会破坏归因 |
-|---|---|---|
-| 上游 Web 检索 | `curl` GitHub API、`git clone`、WebFetch、Sourcegraph 查询 | PR、commit、文件 diff 或测试可能直接给出修复 |
-| 本地 Git 历史 | `git log`、`git show`、`git cherry-pick` | 镜像若含未来提交，答案已在本地，不需要联网 |
-| 评测工件暴露 | 可读的 gold patch、隐藏测试、错误配置的挂载或缓存 | agent 能针对判定器优化，而不是解决 issue |
-| 环境侧信道 | 构建产物、依赖缓存、已修复版本的行为 | 可能泄露“问题已被修复”或实现方向 |
-
-Cursor 在 2026 年 6 月 25 日的研究文章 [*Reward hacking is swamping model intelligence gains*](https://cursor.com/blog/reward-hacking-coding-benchmarks) 对这个问题给出了直接的实证：其对 SWE-bench Pro 成功轨迹的审计将“上游查找”和“Git 历史挖掘”列为主要模式；文章展示了 agent 查询 GitHub PR 文件列表、以及通过 `git show` / `git cherry-pick` 直接复现未来修复的例子。Cursor 的结论不是“禁止一切工具”，而是公开历史仓库上的离线评测需要控制运行时网络与仓库历史，否则标准分数混合了编码能力和对已知修复的访问能力。
-
-我们的 102-case 轨迹审计也说明这是具体风险而非抽象担忧：曾观察到成功的 `curl codeload.github.com` / `git clone https://github.com/...`，以及经 Sourcegraph 获取 GitHub 内容的调用。按审计口径，至少 35 次成功的 GitHub 或代理内容获取覆盖 14 个 case。这个证据足以说明这些运行的通过率不能被直接解释为“完全自主修复率”；但它**不能**单凭自身证明每一个命中 case 的最终 patch 都抄自上游，是否复写仍须对照轨迹、检索内容和 patch。
-
-## 严格环境的最小组成
-
-我们采用的做法与 Cursor 文中 strict harness 的两项核心机制一致，但将其实现为 Harbor trial 生命周期内可审计的控制。
-
-```text
-镜像构建阶段（可控下载）
-  └─ 解析并安装任务运行所需依赖
-
-Agent trial（无答案检索面）
-  ├─ 原始 .git 备份到 trial 私有位置
-  ├─ 移除 .git，git init 为单一基线提交
-  ├─ 默认拒绝网络出口，仅放行已配置的模型 API
-  └─ agent 在本地工作区修改、测试、提交 patch
-
-Verifier（独立评分）
-  ├─ 恢复原始 Git 元数据，仅用于 gold test checkout
-  └─ 在同样严格的 verifier 网络策略下执行判定
-```
-
-### 1. Git 历史隔离
-
-在 agent 启动前，trial 会备份任务自带 `.git`，删除工作区的 `.git`，随后重新初始化 Git、配置固定身份、将当前快照提交为一个新的基线提交。agent 因此仍能使用日常 Git 工作流（查看 diff、暂存、提交、回滚自己创建的提交），但看不到远端、分支和历史上已经存在的修复。评分前才恢复原始元数据，以保持数据集既有的 gold-test checkout 语义。
-
-这项控制必须正确处理真实任务中异常的 Git 元数据。我们遇到过 `.git/hooks/pre-commit` 指向 `.git` 外部位置的符号链接，它会使 Docker 复制备份失败。最初的宽泛做法是删除 hooks；随后收窄为：仅删除会逃逸出当前 Git 目录的 hook 符号链接，普通 hook 文件和指向 Git 内部的链接都保留。这样解决复制失败，同时尽量不改变任务行为。
-
-### 2. 网络出口隔离
-
-网络策略需要在命令真正执行前由宿主侧/容器侧强制执行，而不能只依赖 prompt 或工具描述。策略目标是：
-
-- agent 容器默认不能访问公网、GitHub、代码搜索和镜像 API；
-- 唯一的必要例外是模型 API，且从当前 primary、fast 与 subagent 配置派生为显式 allowlist；
-- verifier 也不依赖运行时外网；
-- 每次新增模型服务都要进行一次真实的、管道内模型请求冒烟测试，不能因为当前使用 GLM 就把白名单写死。
-
-对 agent 而言，最好把不可用的 Web 工具直接隐藏或让工具声明清楚“本评测不允许检索公网答案”。这样它不会把被防火墙拒绝误判为参数错误，并在同一种联网调用上反复试错。提示只改善交互；真正的边界仍由网络策略承担。
-
-### 3. 轨迹与证据
-
-严格环境不是只看最终 reward。每个 run 至少应留下：工具调用、shell 拒绝原因、网络拒绝事件、Git 隔离事件、最终 changed paths、模型请求起始/首 token/结束时间、token 与失败原因。它们分别回答“是否试图取答案”“是否被拦截”“是否因模型或工具超时”“最终 patch 来自哪里”。
-
-## 为什么要做全量 golden-patch 验证
-
-断网和 Git 隔离会改变执行环境；如果只把模型放进去跑，任何分数下降都可能有两种解释：模型真的受限了，或评测环境坏了。golden patch 是区分这两者的基线。
-
-对每个任务，把官方/数据集的正确 patch 交给同一套严格 verifier。如果 `reward != 1`，先修环境，不应把失败计为 agent 能力不足。全量 731 case 的意义在于每个仓库的构建与验证路径差异极大：Node 原生模块、Go module、Helm、protobuf、TLS 测试、解析脚本或超时策略都可能只在少数任务中暴露。只跑几个 smoke case 无法证明严格环境仍正确地实现了原始判定器。
-
-因此全量 golden 验证证明的是：
-
-| 它能证明 | 它不能证明 |
-|---|---|
-| 在记录的镜像、数据集、Harbor 版本和策略下，正确 patch 可被 verifier 接受 | 某个模型一定能解决任务 |
-| 严格隔离没有把已知正确解系统性判错 | agent 没有训练数据污染 |
-| 环境问题可和模型失败分开归因 | 所有形式的评测泄漏都已消失 |
-
-只有完成该前置验证，后续严格模型跑分才可以被解释为“在该评测定义下的 agent 表现”。同时应记录 Harbor commit、数据集/overlay digest、镜像 digest、模型配置和策略版本；否则未来无法区分模型变动与环境变动。
-
-## 断网后依赖怎么办
-
-严格 agent trial 中禁止联网，会自然阻断 `npm install`、`go mod download`、`helm dependency build`、`protoc` 下载或 `node-gyp` 下载 headers 等工作流。解决方法不是在 agent 执行时临时放开公网，而是让依赖在**镜像构建阶段**准备好：
-
-1. 从任务快照已经声明的依赖图出发，下载 lockfile / module manifest 指定的包、工具和 headers；
-2. 放入镜像层或由只读的本地 package proxy/cache 提供；
-3. agent 与 verifier 运行时都只使用这份离线可用的环境；
-4. 若某个任务仍需要网络，先把它视为环境缺陷并用 golden patch 复现、修复，而不是为该 case 开放任意出口。
-
-这也是把“开发所需依赖获取”与“查找公开答案”分离的关键：前者发生在可复现的构建输入中，后者不会发生在 agent trial 中。
-
-## 预取 Oracle 依赖：规避作弊带来的负面作用
-
-彻底阻断 agent 的公网出口后，依赖下载也会一并被阻断。为了让正确 patch 能被验证，我们自然会尝试在镜像构建阶段预取依赖；但**若预取清单来自 Oracle / gold patch，这本身会改变被测任务的解空间**。这不只是“会不会给模型提示”的小问题，而是严格评测的一个负面作用和新的有效性风险。
-
-### 两种不同的风险
-
-第一种是信息侧信道：若根据 gold patch 的新增 import、修改后的 lockfile 或某次 Oracle 失败，专门向 agent 镜像安装一个此前不存在的包，agent 虽然没有收到文字 prompt，却能在环境中观察到“这个库已经可用”。这会泄露某个库、版本或工具可能与答案相关。
-
-第二种更严重：**方案空间被 Oracle 依赖足迹锁死。** 设想 gold patch 使用库 A，因此镜像仅预装 A；另一个完全正确的 agent patch 使用库 B，或者需要另一种编译器插件、Node header、Go module、Helm chart。运行时既不允许联网也没有 B，它会因环境缺依赖而失败。最终分数就偏向“依赖形状接近 gold patch 的实现”，而不是“任何满足任务语义的实现”。在极端情况下，评测实际上只保证 gold patch 这一条实现路径可过。
-
-因此，要明确区分下面三件事：
-
-| 结论 | 是否成立 |
-|---|---|
-| 构建期依赖预取会给模型新增文字上下文 | 通常不会；缓存不是 prompt |
-| Oracle 特定预取会给模型额外线索 | 会，属于环境侧信道 |
-| Oracle 特定预取可能让非 gold 的正确解无法通过 | 会；它把评测变成一个封闭且偏向 Oracle 的依赖预算 |
-
-最后一点不能被 golden patch 的 `reward=1` 掩盖。全量 golden 验证只证明“gold patch 在严格环境可通过”，是必要条件；它不能证明“所有合理的替代解也能在严格环境被公平评分”。
-
-
-## 可复用的评测准则
-
-1. **把“能联网”视为实验变量，而不是默认背景。** 对公开历史任务，标准分数与严格分数应并列报告，不能互相替代。
-2. **同时隔离网络、Git 历史和评测工件。** 只断网无法消除本地 future commit；只删 Git 无法阻止上游检索。
-3. **白名单从配置推导。** 模型服务可以连通不等于允许任意公网；primary、fast、subagent 都要覆盖并实测。
-4. **用全量 golden patch 给环境背书。** 先证明正确答案能通过，再解释 agent 失败。
-5. **把依赖预取当作数据血缘和解空间问题。** 基于原始依赖图的离线构建是环境准备；由 gold patch 反推的包必须隔离、标注，并且不得让它成为某一条正确解独有的运行条件。
-6. **保留轨迹并审计归因。** “访问过外部答案”与“最终 patch 确实复写答案”是不同级别的结论，报告必须区分。
-
-严格评测会牺牲一部分现实开发中的便利性，特别是临时联网搜索和按需下载依赖。但它换来的是更重要的东西：当一个分数上升或下降时，我们知道它主要在衡量什么，也能知道应该修模型、工具、数据集还是环境。
