+++ src/crates/adapters/ai-adapters/src/providers/gemini/request.rs
@@ -348,7 +348,6 @@ pub(crate) async fn send_stream(
 
     execute_sse_request(
         "Gemini Streaming API",
-        "gemini",
         &url,
         &request_body,
         max_tries,
