+++ src/crates/adapters/ai-adapters/src/providers/openai/codex_chatgpt.rs
@@ -172,7 +172,6 @@ pub(crate) async fn send_stream(
 
     execute_sse_request(
         "Codex ChatGPT Responses API",
-        "codex_chatgpt",
         &url,
         &request_body,
         max_tries,
