+++ src/crates/assembly/core/src/agentic/tools/implementations/shell_command_safety.rs
@@ -44,51 +44,11 @@ const MUTATING_COMMANDS: &[&str] = &[
     "sed", "tee", "touch", "truncate", "unzip",
 ];
 
-const BLOCKED_EVALUATION_CODE_SOURCE_DOMAINS: &[&str] = &[
-    "github.com",
-    "githubusercontent.com",
-    "githubassets.com",
-    "git.io",
-    "sourcegraph.com",
-];
-
-const NETWORK_GIT_SUBCOMMANDS: &[&str] =
-    &["clone", "fetch", "pull", "push", "ls-remote", "submodule"];
-
-const REVISION_READING_GIT_SUBCOMMANDS: &[&str] = &[
-    "archive",
-    "blame",
-    "cat-file",
-    "checkout",
-    "diff",
-    "format-patch",
-    "grep",
-    "log",
-    "range-diff",
-    "reflog",
-    "restore",
-    "rev-list",
-    "show",
-    "switch",
-    "whatchanged",
-];
-
 pub(crate) fn exec_command_is_concurrency_safe(command: &str) -> bool {
     classify_command(command) == ShellCommandRisk::ReadOnly
 }
 
 pub(crate) fn denial_for_command(command: &str) -> Option<String> {
-    if contains_blocked_evaluation_code_source_domain(command) {
-        return Some(
-            "Refused in this evaluation: shell commands cannot access GitHub or source-code mirrors such as Sourcegraph. Inspect only the local repository and task materials."
-                .to_string(),
-        );
-    }
-
-    if let Some(message) = evaluation_git_history_denial(command) {
-        return Some(message);
-    }
-
     if opens_original_app_sqlite_db_with_invalid_wal(command) {
         return Some(
             "Refused: opening a SQLite database whose sibling WAL has an invalid header can \
@@ -99,153 +59,6 @@ pub(crate) fn denial_for_command(command: &str) -> Option<String> {
     }
 
     None
-}
-
-fn contains_blocked_evaluation_code_source_domain(command: &str) -> bool {
-    command
-        .to_ascii_lowercase()
-        .split(|character: char| {
-            !(character.is_ascii_alphanumeric() || matches!(character, '.' | '-'))
-        })
-        .filter(|token| !token.is_empty())
-        .any(|token| {
-            BLOCKED_EVALUATION_CODE_SOURCE_DOMAINS.iter().any(|domain| {
-                token == *domain
-                    || token
-                        .strip_suffix(domain)
-                        .is_some_and(|prefix| prefix.ends_with('.'))
-            })
-        })
-}
-
-fn evaluation_git_history_denial(command: &str) -> Option<String> {
-    for segment in command_segments(command) {
-        if contains_command_named(segment, &["gh", "hub"]) {
-            return Some(
-                "Refused in this evaluation: GitHub command-line clients are disabled because they can retrieve an upstream solution. Use only the repository already present in the workspace."
-                    .to_string(),
-            );
-        }
-
-        let Some((subcommand, arguments)) = git_subcommand_and_arguments(segment) else {
-            continue;
-        };
-
-        if NETWORK_GIT_SUBCOMMANDS.contains(&subcommand.as_str())
-            || (subcommand == "remote"
-                && arguments
-                    .first()
-                    .is_some_and(|argument| argument == "update"))
-            || (subcommand == "lfs"
-                && arguments
-                    .first()
-                    .is_some_and(|argument| matches!(argument.as_str(), "fetch" | "pull")))
-        {
-            return Some(
-                "Refused in this evaluation: Git network operations are disabled because they can retrieve an upstream solution. Use only the repository already present in the workspace."
-                    .to_string(),
-            );
-        }
-
-        if REVISION_READING_GIT_SUBCOMMANDS.contains(&subcommand.as_str())
-            && arguments
-                .iter()
-                .take_while(|argument| argument.as_str() != "--")
-                .any(|argument| is_forbidden_evaluation_revision(argument))
-        {
-            return Some(
-                "Refused in this evaluation: remote-tracking refs, reflogs, and explicit commit hashes may expose the benchmark target patch. Inspect the current checkout and working-tree diff instead."
-                    .to_string(),
-            );
-        }
-    }
-    None
-}
-
-fn contains_command_named(segment: &str, names: &[&str]) -> bool {
-    segment.split_whitespace().any(|token| {
-        let token = token.trim_matches(|character: char| {
-            matches!(character, '(' | ')' | '[' | ']' | '{' | '}' | '\'' | '"')
-        });
-        Path::new(token)
-            .file_name()
-            .and_then(|name| name.to_str())
-            .is_some_and(|name| {
-                names
-                    .iter()
-                    .any(|candidate| name.eq_ignore_ascii_case(candidate))
-            })
-    })
-}
-
-fn git_subcommand_and_arguments(segment: &str) -> Option<(String, Vec<String>)> {
-    let tokens = segment
-        .split_whitespace()
-        .map(|token| {
-            token
-                .trim_matches(|character: char| {
-                    matches!(character, '(' | ')' | '[' | ']' | '{' | '}' | '\'' | '"')
-                })
-                .to_string()
-        })
-        .filter(|token| !token.is_empty())
-        .collect::<Vec<_>>();
-    let git_index = tokens.iter().position(|token| {
-        Path::new(token)
-            .file_name()
-            .and_then(|name| name.to_str())
-            .is_some_and(|name| name.eq_ignore_ascii_case("git"))
-    })?;
-
-    let mut index = git_index + 1;
-    while index < tokens.len() {
-        let token = &tokens[index];
-        if matches!(
-            token.as_str(),
-            "-C" | "-c" | "--git-dir" | "--work-tree" | "--namespace"
-        ) {
-            index += 2;
-            continue;
-        }
-        if token.starts_with('-') {
-            index += 1;
-            continue;
-        }
-
-        return Some((
-            token.to_ascii_lowercase(),
-            tokens[index + 1..]
-                .iter()
-                .map(|argument| argument.to_ascii_lowercase())
-                .collect(),
-        ));
-    }
-    None
-}
-
-fn is_forbidden_evaluation_revision(argument: &str) -> bool {
-    let argument = argument.trim_matches(|character: char| {
-        matches!(
-            character,
-            '\'' | '"' | '(' | ')' | '[' | ']' | '{' | '}' | ','
-        )
-    });
-    argument == "--all"
-        || argument == "--reflog"
-        || argument == "fetch_head"
-        || argument == "orig_head"
-        || argument.contains("origin/")
-        || argument.contains("upstream/")
-        || argument.contains("refs/remotes/")
-        || argument.contains("remotes/")
-        || argument.contains("@{u")
-        || contains_explicit_object_id(argument)
-}
-
-fn contains_explicit_object_id(argument: &str) -> bool {
-    argument
-        .split(|character: char| !character.is_ascii_hexdigit())
-        .any(|part| (7..=40).contains(&part.len()))
 }
 
 fn classify_command(command: &str) -> ShellCommandRisk {
@@ -410,8 +223,8 @@ fn sqlite_wal_has_invalid_header(db_path: &Path) -> bool {
 #[cfg(test)]
 mod tests {
     use super::{
-        classify_command, contains_blocked_evaluation_code_source_domain, denial_for_command,
-        exec_command_is_concurrency_safe, extract_original_app_db_paths, ShellCommandRisk,
+        classify_command, denial_for_command, exec_command_is_concurrency_safe,
+        extract_original_app_db_paths, ShellCommandRisk,
     };
 
     #[test]
@@ -473,64 +286,5 @@ mod tests {
             "python3 -c \"import sqlite3; sqlite3.connect('/app/main.db')\"",
         )
         .is_none());
-    }
-
-    #[test]
-    fn evaluation_blocks_direct_and_proxied_github_access() {
-        for command in [
-            "curl -L https://codeload.github.com/org/repo/tar.gz/main",
-            "git clone https://github.com/org/repo.git",
-            "wget https://raw.githubusercontent.com/org/repo/main/src/lib.rs",
-            "curl https://sourcegraph.com/github.com/org/repo/-/raw/src/lib.rs",
-            "git clone git@github.com:org/repo.git",
-        ] {
-            assert!(denial_for_command(command)
-                .as_deref()
-                .is_some_and(|message| message.contains("source-code mirrors")));
-        }
-
-        assert!(!contains_blocked_evaluation_code_source_domain(
-            "curl https://notgithub.com/example"
-        ));
-        assert!(denial_for_command("curl https://notgithub.com/example").is_none());
-    }
-
-    #[test]
-    fn evaluation_blocks_git_network_and_hidden_solution_revisions() {
-        for command in [
-            "git fetch origin",
-            "git pull --ff-only",
-            "git push origin HEAD",
-            "git ls-remote upstream",
-            "gh api repos/org/repo/commits/main",
-            "git show origin/main",
-            "git show 878c25b",
-            "git diff HEAD...upstream/main",
-            "git log --all --oneline",
-            "git cat-file -p fd18df1",
-            "git -C /app restore --source=7073d18b src/server.go",
-        ] {
-            assert!(
-                denial_for_command(command).is_some(),
-                "should block: {command}"
-            );
-        }
-    }
-
-    #[test]
-    fn evaluation_keeps_normal_workspace_git_inspection_available() {
-        for command in [
-            "git status --short",
-            "git diff",
-            "git diff --cached",
-            "git diff HEAD -- src/lib.rs",
-            "git log -5 HEAD",
-            "git show HEAD:src/lib.rs",
-        ] {
-            assert!(
-                denial_for_command(command).is_none(),
-                "should allow: {command}"
-            );
-        }
     }
 }
